backend/requirements.txt
--only-binary=all
fastapi>=0.120.0
uvicorn[standard]>=0.31.1
openai>=2.6.0
anthropic==0.7.8
sentence-transformers==2.2.2
huggingface_hub>=0.34.0
pydantic>=2.7.4
python-dotenv>=1.1.0
rich>=13.9.4
httpx==0.27.1
python-multipart>=0.0.9
geopy==2.4.1
requests>=2.32.5
beautifulsoup4>=4.12.0
asyncio>=3.4.3

# Flask dependencies (for event_api)
flask>=3.0.0
flask-limiter>=3.5.0

# Firebase dependencies
firebase-admin>=6.0.0
google-cloud-firestore>=2.11.0
bcrypt>=4.0.0
